+++ bin/rmemo.js
@@ -11,6 +11,7 @@ import { cmdStart } from "../src/cmd/start.js";
 import { cmdDone } from "../src/cmd/done.js";
 import { cmdTodo } from "../src/cmd/todo.js";
 import { cmdTemplate } from "../src/cmd/template.js";
+import { cmdSync } from "../src/cmd/sync.js";
 import { parseArgs, printHelp } from "../src/lib/args.js";
 import { exitWithError } from "../src/lib/io.js";
 
@@ -54,6 +55,9 @@ try {
       break;
     case "template":
       await cmdTemplate({ rest, flags });
+      break;
+    case "sync":
+      await cmdSync({ flags });
       break;
     case "help":
     case undefined: